project1/flask/bitly/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
nanoid==2.0.0
